feat: refactor AiChatTab import paths and relocate component to agent directory
