+++ src/renderer/components/states/maxExpand/components/agent/index.ts
@@ -24,4 +24,4 @@
  * @author 鸡哥
  */
 
-export { AiChatTab } from '../AiChatTab';
+export { AiChatTab } from './components/AiChatTab';
